Implement new coordinate storage system
All coordinates are now stored as a single unsigned 32-bit integer. The
first (upper) 16 bits of the integer are used for storing the X value
while the other (lower) 16 bits are used for storing the Y value. In
turn, this allows me to refactor drawRectangle to use all the available
argument registers.

It's possible that we can store width and height the same way to, which
will allow me to refactor drawRectangle to use only 3 argument
registers, but I don't see a good relationship between the two to
justify it. Plus, I don't want to potentially confuse anyone more than I
already have to.