Add drawLine routine
This routine will be used for a new drawNumber routine later. However,
it won't be used in drawRectangle for performance reasons. We can
rewrite drawRectangle to be a glorified loop that calls drawLine over
and over with readjusting which line to draw next to, but that would add
a bit of additional overhead.