Optimize drawRectangle routine
Instead of keeping a counter for the outer loop, find the ending address
of the entire rectangle instead and compare it with our ending address
of the current row. This eliminates one instruction from the loop, but
in turn we place another instruction outside of the loop. I will still
take this as a slight optimization, however.